requirements.txt
Django>=5.0
Pillow>=10.0
python-pptx>=1.0.2
pywin32>=306
